Only allow OH owners to delete queue entries
TODO: Update rendering logic to render queue entry cards in rails
and only render the delete button if the user can delete the card.

Delete should be enabled for the OH owner and the student that
created the card, even if they aren't logged in. This can be
accomplished via a session cookie.