Enforce authorization to update / delete OH and queue entries (#10)
* Add misc dev files to .gitignore

* Specify migration version for rails 6 compatibility

* Enforce user ownership of office hours

- Enforce authentication directly with devise for the following OH
  routes: new, create, update, edit, destroy

- Add a non-nullable FK to office hours referencing the users table
- Users cannot delete or edit office hours they didn't create
  (don't own). This is enforced at the controller level.
- Don't render edit / delete buttons for office hours the user
  doesn't own.
- Add an optional FK to queue_entries referencing users

* Streamline office hours show view

* Only allow OH owners to delete queue entries

TODO: Update rendering logic to render queue entry cards in rails
and only render the delete button if the user can delete the card.

Delete should be enabled for the OH owner and the student that
created the card, even if they aren't logged in. This can be
accomplished via a session cookie.